requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
aspy.yaml==1.3.0
attrs==23.2.0
Automat==22.10.0
autopep8==2.3.2
backtest==0.7.8
Backtesting==0.3.3
beautifulsoup4==4.12.3
black==19.3b0
blis==1.2.0
bokeh==3.4.1
boto3==1.34.108
botocore==1.34.108
bs4==0.0.2
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
chromedrivermanager==0.0.1
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
colored==2.2.4
confection==0.1.5
configparse==0.1.5
configparser==7.0.0
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.11
datasets==3.2.0
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dnspython==2.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
fake-useragent==0.1.11
fastapi==0.115.6
fastapi-pagination==0.12.34
filelock==3.14.0
fonttools==4.51.0
frozendict==2.4.1
frozenlist==1.5.0
fsspec==2024.9.0
future==0.16.0
fuzzywuzzy==0.18.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.27.1
hyperlink==21.0.0
identify==2.5.36
idna==3.7
importlib_metadata==7.1.0
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.26.1
logging==0.4.9.6
lxml==5.2.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
MetaTrader5==5.0.4288
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
murmurhash==1.0.12
networkx==3.4.2
nltk==3.5
nodeenv==1.8.0
numpy==1.26.4
openai==1.28.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pandas_ta==0.3.14b0
parrot @ git+https://github.com/PrithivirajDamodaran/Parrot_Paraphraser.git@03084c54b64019ba5fa0b620b9c70ad81123e458
parsel==1.9.1
peewee==3.17.1
pillow==10.3.0
platformdirs==4.2.2
plotly==5.21.0
pluggy==1.5.0
pre-commit==1.16.1
preshed==3.0.9
propcache==0.2.1
Protego==0.3.1
proxyscrape==0.3.0
psutil==6.1.1
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PyDispatcher==2.0.7
PyGetWindow==0.0.9
Pygments==2.17.2
Pymem==1.14.0
pymongo==4.10.1
PyMsgBox==1.0.9
pynput==1.7.7
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
queuelib==1.7.0
random_word==1.0.13
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
s3transfer==0.10.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
Scrapy==2.11.2
scrapy-proxy-pool==0.1.9
scrapy-user-agents==0.1.1
secret==0.8
selenium==4.21.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-wordnet==0.1.0
spark-parser==1.8.9
srsly==2.5.1
starlette==0.41.3
sympy==1.13.1
ta==0.11.0
tabulate==0.9.0
tenacity==8.2.3
term-background==1.0.1
termcolor==2.4.0
thinc==8.3.4
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4
tqdm==4.67.1
transformers==4.47.1
trio==0.25.1
trio-websocket==0.11.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
uncompyle6==3.9.1
urllib3==2.2.1
user-agents==2.2.0
uvicorn==0.34.0
virtualenv==20.26.2
w3lib==2.2.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webdriver-manager==4.0.1
webencodings==0.5.1
wrapt==1.17.2
wsproto==1.2.0
xdis==6.1.0
xxhash==3.5.0
xyzservices==2024.4.0
yarl==1.18.3
yfinance==0.2.37
you-get @ file:///C:/Users/ahmet/OneDrive/Desktop/testing/you-get
zipp==3.18.2
zope.interface==6.4.post2
